+++ package.json
@@ -6,7 +6,7 @@
   "private": true,
   "scripts": {
     "build": "node build.js",
-    "dev": "wrangler dev",
+    "dev": "wrangler pages dev --kv storage --proxy 3000 --kv-persist ./miniflare-kv -- serve static",
     "format": "prettier --write '*.{json,js}' 'src/**/*.{js,jsx,ts,tsx}'",
     "lint": "eslint src --max-warnings=0 && prettier --check '*.{json,js}' 'src/**/*.{js,jsx,ts,tsx}'",
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -31,6 +31,7 @@
     "esbuild": "^0.14.25",
     "eslint": "^8.9.0",
     "eslint-config-prettier": "^8.5.0",
+    "miniflare": "^2.4.0",
     "moviedb-promise": "^3.1.13",
     "prettier": "^2.5.1",
     "typescript": "^4.5.5"
